refactor: routing with state, not react-router
Why not use react-router? Since we're developing a game, we don't want
users to navigate to different screens by changing the URL in the
address bar, e.g., going to `/howtoplay` to access the game mechanics.

Rather, we want the game to be accessible only in the root path. This
implementation is closer to how an actual game works, i.e., the game
always starts in the main menu.